Add UIKit import
As OHAlertView inherits from UIAlertView, it needs UIKit to be imported to compile. Right now it's made
by putting UIKit into precompiled header for the project. But newer versions of XCode does not create a
precompiled header for the project anymore by default, therefore it won't compile as UIAlertView class is
not defined. Solved the problem by adding "import UIKit" statement in header file itself